refactor: Use fake language loader to load earlier
* This allows Debugify (if loaded by Sinytra Connector) to correctly load its configs as otherwise Debugify config checker will load earlier. This will also fix config copying for any mods that disable mixins with mixin plugin like Debugify.
