+++ src/core/resources/META-INF/services/net.neoforged.neoforgespi.language.IModLanguageLoader
@@ -1,0 +1,1 @@
+me.shedaniel.yosbr.core.YOSBRLanguageLoader
