--- src/main/java/me/shedaniel/yosbr/mixin/YOSBRMixinPlugin.java
@@ -1,24 +1,0 @@
-package me.shedaniel.yosbr.mixin;
-
-import me.shedaniel.yosbr.YourOptionsShallBeRespected;
-import org.objectweb.asm.tree.ClassNode;
-import org.spongepowered.asm.mixin.extensibility.IMixinConfigPlugin;
-import org.spongepowered.asm.mixin.extensibility.IMixinInfo;
-
-import java.util.List;
-import java.util.Set;
-
-// This is a MESSED UP hack we do to load before mods initialize their configs...
-public class YOSBRMixinPlugin implements IMixinConfigPlugin {
-    @Override
-    public void onLoad(String mixinPackage) {
-        YourOptionsShallBeRespected.respectTheOptions();
-    }
-
-    @Override public String getRefMapperConfig() { return ""; }
-    @Override public boolean shouldApplyMixin(String targetClassName, String mixinClassName) { return true; }
-    @Override public void acceptTargets(Set<String> myTargets, Set<String> otherTargets) {}
-    @Override public List<String> getMixins() { return List.of(); }
-    @Override public void preApply(String targetClassName, ClassNode targetClass, String mixinClassName, IMixinInfo mixinInfo) {}
-    @Override public void postApply(String targetClassName, ClassNode targetClass, String mixinClassName, IMixinInfo mixinInfo) {}
-}
