--- src/main/resources/yosbr.mixins.json
@@ -1,10 +1,0 @@
-{
-  "required": true,
-  "package": "me.shedaniel.yosbr.mixin",
-  "plugin": "me.shedaniel.yosbr.mixin.YOSBRMixinPlugin",
-  "compatibilityLevel": "JAVA_21",
-  "refmap": "yosbr.refmap.json",
-  "mixins": [],
-  "client": [],
-  "minVersion": "0.8"
-}
